Switch to using types from the http crate
This change marks a first step towards decoupling the crate from hyper.
We want the crate to focus on providing a way to describe what an
endpoint looks like and be as independent as possible from the entity
actually issuing a request. With this patch, specifically, we switch
over to using types from the http crate over the re-exports from hyper.
The http crate is considered more neutral than hyper and surely will
make it easier to transition to custom types should the need for doing
so arise in the future.