Preserve message body in HttpStatus error variant
A lot of HTTP endpoints provide additional information pertaining an
error inside the body of the response. So far we have ignored this data
when converting an endpoint specific error into our more generic one.
With this change we preserve this data as good as we can. Note that the
approach we use is not quite as clean as we would like it to be. In
particular, if the error was a "known" variant then what we end up
preserving is the textual representation of the parsed "API error",
which got produced from the response body, not the actual body.