Require Debug, Display, and Error for Endpoint::Error
So far we have been pretty loose with the requirements for the
individual associated types in the Endpoint trait. That was mainly done
to keep things simple and because requirements did not dictate
otherwise.
Unfortunately, especially the loosely defined Error type in this trait
is causing interoperability problems, as commonly used trait
implementations are missing.
With this change we require that the type used as an error implements
Debug, Display, and Error from the standard library, in addition to the
two conversion already present previously.