Do not let CI code coverage artifacts expire
As part of every CI run we collect code coverage artifacts in the form
of HTML pages. From the README we link to these pages to have a
graphical representation.
GitLab has decided to remove those artifacts after a certain time,
rendering our link to them useless.
With this change we remove the expiration date such that these artifacts
are no longer deleted behind our back.