Require Error instead of Display implementation for Endpoint::ApiError
The Endpoint::ApiError is an associated type representing an error. Yet,
for the sake of simplicity we only required that it implements Display
and not the standard Error trait.
For the sake of uniformity, with this change we switch over to just
having it require an implementation of the standard Error trait.

On the testing side we now need to create a "proper" error. However,
given that the API we work with does not actually return any error
information, this is just a dummy type.