Make Endpoint::body return an optional body
Depending on the type of HTTP request being performed, an empty body may
be fine. So far we have encoded such an empty body just as an empty
array of bytes. However, it arguably makes more sense to make the
optionality more explicit by expressing that fact via an Option type.
With this change we go down that route.