Remove JSON specific logic
We intentionally have assumed JSON conversion of object and, hence, kept
JSON specific details in the crate, because at the time no APIs we were
interfacing with returned anything else and there were some usability
savings to be had.
With this change we move away from this approach. Specifically, we
remove all JSON specifics from non-test code. Instead of having JSON
error variants, we introduce and use a more or less generic conversion
error.
One negative (but necessary) side effect of the removal of default
implementations is that clients now need to provide their own parsing
methods -- on each endpoint definition. The assumption here is that
clients can more or less easily wrap the existing macro and provide
default parsing this way.