Make Endpont::{body, parse} return more specific errors
So far the Endpont::{body, parse} methods were returning the endpoint's
error type in case of a problem. That's arguably too much flexibility:
it includes HTTP error variants that should never be possible to create
in this context.
Given that with the Endpoint::ConversionError we now have an error type
suitable for exactly these kind of contexts, this change adjusts the
method signatures to use it.