Switch from test-env-log to test-log
The test-env-log crate has recently been renamed to test-log in order to
better reflect its intended purpose. With this change we switch over to
using test-log in our tests accordingly.
While at it, also enable its 'trace' feature to initialize the tracing
crate, as tracing output is actually what is interesting to look at when
running the tests.